+++ packages/browser-wallet/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+-   Added support for two additional types of Sponsored Transactions, `Configure Delegation` and `Update Smart Contract Instance`.
 -   Added export activity button to Transaction Log as part of feature parity with Old UI
 -   Added Allowlist config at account page as part of feature parity with Old UI